Added comments throughout the codebase.
Upon inspecting the codebase, I realized there was very limited commenting, and
what commenting there was was only suitable for my own consumption. So I went
through a first pass of adding comments to torus, as it's rather complex, and
the more commonly used functions. I also added more whitespace to trig.scad
to enhance legibility.

This is no substitute for proper documentation, but hopefully will serve in
the short term.